fix: move ContentReveal to shared components and migrate profile-card onto it
