+++ clients/web/src/components/content-reveal.tsx
