+++ clients/web/src/domains/settings/components/daily-credit-limit-card.tsx
@@ -1,7 +1,7 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useState, type ChangeEvent } from "react";
 
-import { ContentReveal } from "@/domains/settings/components/content-reveal";
+import { ContentReveal } from "@/components/content-reveal";
 import { SkeletonLines } from "@/domains/settings/components/skeleton-lines";
 import { extractDrfFieldErrors } from "@/domains/settings/utils/drf-errors";
 import {
