+++ clients/web/src/domains/settings/components/payment-methods-card.tsx
@@ -9,7 +9,7 @@ import { Skeleton } from "@vellumai/design-library/components/skeleton";
 
 import { AutoTopUpPaymentMethodModal } from "@/domains/settings/components/auto-top-up-payment-method-modal";
 import { BillingSectionHeader } from "@/domains/settings/components/billing-section-header";
-import { ContentReveal } from "@/domains/settings/components/content-reveal";
+import { ContentReveal } from "@/components/content-reveal";
 import { SkeletonLines } from "@/domains/settings/components/skeleton-lines";
 import {
   modalSnapshotFor,
